connected all the button to the switch panel listener/implement actionListener for switching the listeners instead of mouseListener

+++ model/RightFlipper.java
@@ -116,12 +116,12 @@ public class RightFlipper implements Flipper{
     }
 
     @Override
-    public int getXPivot() {
+    public double getXPivot() {
         return xPos;
     }
 
     @Override
-    public int getYPivot() {
+    public double getYPivot() {
         return yPos;
     }
